api-gateway/requirements.txt
stapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pyjwt[crypto]==2.8.0
python-multipart==0.0.6